backend/requirements.txt
flask
flask-cors
resend
python-dotenv
flask-limiter
twilio
